+++ src/commonMain/kotlin/exchange.dydx.abacus/processor/wallet/account/OrdersProcessor.kt
@@ -1,8 +1,8 @@
 package exchange.dydx.abacus.processor.wallet.account
 
-import BlockAndTime
 import exchange.dydx.abacus.processor.base.BaseProcessor
 import exchange.dydx.abacus.protocols.ParserProtocol
+import exchange.dydx.abacus.state.manager.BlockAndTime
 import exchange.dydx.abacus.utils.mutable
 import exchange.dydx.abacus.utils.typedSafeSet
