+++ src/commonMain/kotlin/exchange.dydx.abacus/processor/wallet/account/deprecated/V3AccountProcessor.kt
@@ -1,9 +1,9 @@
 package exchange.dydx.abacus.processor.wallet.account.deprecated
 
-import BlockAndTime
 import exchange.dydx.abacus.processor.base.BaseProcessor
 import exchange.dydx.abacus.processor.wallet.account.SubaccountProcessor
 import exchange.dydx.abacus.protocols.ParserProtocol
+import exchange.dydx.abacus.state.manager.BlockAndTime
 import exchange.dydx.abacus.utils.mutable
 import exchange.dydx.abacus.utils.safeSet
