+++ src/commonMain/kotlin/exchange.dydx.abacus/state/manager/Payloads.kt
@@ -1,4 +1,4 @@
-package exchange.dydx.abacus.state.manager.utils
+package exchange.dydx.abacus.state.manager
 
 import kollections.JsExport
 import kotlinx.serialization.Serializable
