+++ src/commonMain/kotlin/exchange.dydx.abacus/state/model/TradingStateMachine+Wallet.kt
@@ -1,9 +1,9 @@
 package exchange.dydx.abacus.state.model
 
-import BlockAndTime
 import exchange.dydx.abacus.responses.SocketInfo
 import exchange.dydx.abacus.state.changes.Changes
 import exchange.dydx.abacus.state.changes.StateChanges
+import exchange.dydx.abacus.state.manager.BlockAndTime
 import kollections.iListOf
 import kollections.iMutableListOf
 import kollections.toIList
