+++ src/commonTest/kotlin/exchange.dydx.abacus/payload/v4/V4AccountTests.kt
@@ -1,9 +1,9 @@
 package exchange.dydx.abacus.payload.v4
 
-import BlockAndTime
 import exchange.dydx.abacus.payload.BaseTests
 import exchange.dydx.abacus.responses.StateResponse
 import exchange.dydx.abacus.state.app.adaptors.AbUrl
+import exchange.dydx.abacus.state.manager.BlockAndTime
 import exchange.dydx.abacus.state.manager.NotificationsProvider
 import exchange.dydx.abacus.state.model.historicalTradingRewards
 import exchange.dydx.abacus.state.model.onChainAccountBalances
